Break REST API tests out into separate files
one per controller, so that we can track these more effectively.
